requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
ray[default]
tensordict<=0.6.2
# torchdata==1.1.0
torchdata
transformers
vllm==0.8.5
wandb
